backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "university-chatbot-api"
version = "0.1.0"
description = "Multi-tenant university website RAG API (FastAPI)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13",
    "celery[redis]>=5.3",
    "fastapi>=0.115",
    "structlog>=24.0",
    "httpx>=0.27",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.9",
    "redis>=5.0",
    "sqlalchemy>=2.0",
    "uvicorn[standard]>=0.32",
    "psycopg[binary]>=3.2",
    "pgvector>=0.3.5",
    "bcrypt>=4.0",
    "PyJWT>=2.9",
    "email-validator>=2.0",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
app = ["py.typed"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "tests"]
exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"app/models/enums.py" = ["UP042"]
"app/auth/tokens.py" = ["UP017"]
"app/services/auth_service.py" = ["UP017"]
"app/crawler/types.py" = ["UP042"]
"app/services/crawl_db_sink.py" = ["UP017"]
"app/services/crawl_execution_service.py" = ["UP017"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q --strict-markers"
markers = [
    "integration: tests that require Postgres, Redis, or other external services",
]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
omit = ["*/tests/*", "*/alembic/*"]
